bids EP moved
