+++ talentmap_api/fsbid/urls/bidlist.py
@@ -10,7 +10,7 @@ urlpatterns = [
     url(r'^position/(?P<pk>[0-9]+)/$', views.FSBidListPositionActionView.as_view(), name='bidding.FSBid-position-actions'),
     url(r'^position/(?P<pk>[0-9]+)/submit/$', views.FSBidListBidActionView.as_view(), name='bidding.FSBid-bid-actions'),
     url(r'^$', views.FSBidListView.as_view(), name="bidding-FSBid-bid-actions"),
-    #     here is another good place for /bids
+    url(r'^bids/(?P<pk>[0-9]+)/$', views.BidsView.as_view(), name='bidding.bids'),
 ]
 
 urlpatterns += router.urls
